Refactor: Changed tunnel constructor and handleMessage signature
